Disable queue for unlive story
