+++ src/components/Listen/ListenStoryOverlay.tsx
@@ -33,11 +33,24 @@ const ListenStoryOverlay: React.FC<{ storyFeed: Story[] | undefined }> = ({
       </div>
       <StoryBg image={playerPlaying?.image} />
       <div className="absolute z-10 px-2 py-4 bottom-0 w-full bg-gradient-to-t from-background to-transparent">
-        <p className="text-sm text-foreground-secondary text-center mb-1">
-          {t("listen.promptJoin", { username: user?.username || "" })}
+        <p className="text-lg text-foreground-secondary text-center mb-1">
+          {t(
+            playingStory?.isLive
+              ? "listen.promptJoin"
+              : "listen.promptJoinNolive",
+            {
+              username: user?.username || "",
+            }
+          )}
         </p>
         <Link href={`/story/${playingStory?.id}`}>
-          <a className="btn btn-primary w-full">{t("listen.actionJoin")}</a>
+          <a
+            className={`btn ${
+              playingStory?.isLive ? "btn-primary" : ""
+            } w-full`}
+          >
+            {t("listen.actionJoin")}
+          </a>
         </Link>
       </div>
     </>
