+++ src/components/Story/StoryNav.tsx
@@ -2,8 +2,11 @@ import React, { useMemo } from "react";
 import Link from "next/link";
 import ms from "ms";
 import { useUserQuery, Story } from "~/graphql/gql.gen";
+import { useI18n } from "~/i18n/index";
 
 const StoryNav: React.FC<{ story: Story }> = ({ story }) => {
+  const { t } = useI18n();
+
   const [{ data: { user } = { user: undefined } }] = useUserQuery({
     variables: { id: story.creatorId || "" },
   });
@@ -29,7 +32,13 @@ const StoryNav: React.FC<{ story: Story }> = ({ story }) => {
           <Link href={`/user/${user?.username}`}>
             <a className="font-semibold mr-2">{user?.username}</a>
           </Link>{" "}
-          <span className="text-xs text-foreground-secondary">{dateStr}</span>
+          {story.isLive ? (
+            <span className="font-semibold text-xs bg-pink animate-pulse uppercase leading-none py-0.5 px-1 rounded-full">
+              {t("common.live")}
+            </span>
+          ) : (
+            <span className="text-xs text-foreground-secondary">{dateStr}</span>
+          )}
         </div>
         <div className="text-sm text-foreground-secondary">{story.text}</div>
       </div>
